corrections and adds

+++ ft_bzero.c
@@ -6,16 +6,16 @@
 /*   By: grodrig2 <grodrig2@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/21 18:15:12 by grodrig2          #+#    #+#             */
-/*   Updated: 2025/07/21 19:19:32 by grodrig2         ###   ########.fr       */
+/*   Updated: 2025/07/22 11:27:18 by grodrig2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_bzero(void *s, unsigned long int n)
+void	ft_bzero(void *s, size_t n)
 {
 	unsigned char	*ptr;
-	
+
 	ptr = (unsigned char *)s;
 	while (n > 0)
 	{
